LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/shadows/2d.shadow.attributes.shadowColor.current.basic-expected.txt

2d.shadow.attributes.shadowColor.current.basic
currentColor is computed from the canvas element
Actual output:

PASS currentColor is computed from the canvas element
